backend/requirements.txt
aniso8601==9.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
